# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GolfHero API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /charities:
    get:
      operationId: listCharities
      tags: [charities]
      summary: List all charities
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of charities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Charity"
    post:
      operationId: createCharity
      tags: [charities]
      summary: Create a charity (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharityBody"
      responses:
        "201":
          description: Created charity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charity"
  /charities/{id}:
    get:
      operationId: getCharity
      tags: [charities]
      summary: Get a charity by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Charity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCharity
      tags: [charities]
      summary: Update a charity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharityBody"
      responses:
        "200":
          description: Updated charity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charity"
    delete:
      operationId: deleteCharity
      tags: [charities]
      summary: Delete a charity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /charities/my-selection:
    get:
      operationId: getMyCharitySelection
      tags: [charities]
      summary: Get the current user's selected charity and contribution %
      responses:
        "200":
          description: Charity selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharitySelection"
    put:
      operationId: updateMyCharitySelection
      tags: [charities]
      summary: Update charity selection and contribution percentage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCharitySelectionBody"
      responses:
        "200":
          description: Updated selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharitySelection"

  # ── DRAWS ─────────────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{clerkId}:
    get:
      operationId: getUser
      tags: [admin]
      summary: Get user by Clerk ID (admin)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user profile (admin)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /admin/users/{clerkId}/scores:
    get:
      operationId: adminGetUserScores
      tags: [admin]
      summary: Get user's scores (admin)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        handicap:
          type: ["number", "null"]
        role:
          type: string
          enum: [user, admin]
        charityId:
          type: ["integer", "null"]
        charityContributionPercent:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - email
        - role
        - charityContributionPercent
        - createdAt
...
    Score:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        score:
          type: integer
        scoreDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - score
        - scoreDate
        - createdAt
...
    Charity:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        featured:
          type: boolean
        upcomingEvent:
          type: ["string", "null"]
        upcomingEventDate:
          type: ["string", "null"]
          format: date
        totalReceived:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - featured
        - totalReceived
        - createdAt
    CreateCharityBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        website:
          type: string
        category:
          type: string
        featured:
          type: boolean
        upcomingEvent:
          type: string
        upcomingEventDate:
          type: string
          format: date
      required:
        - name
        - description
    CharitySelection:
      type: object
      properties:
        charityId:
          type: ["integer", "null"]
        charity:
          $ref: "#/components/schemas/Charity"
        contributionPercent:
          type: integer
      required:
        - contributionPercent
    UpdateCharitySelectionBody:
      type: object
      properties:
        charityId:
          type: integer
        contributionPercent:
          type: integer
          minimum: 10
          maximum: 100
      required:
        - charityId
        - contributionPercent
...
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    AdminUpdateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
        firstName:
          type: string
        lastName:
          type: string