# lib/db/src/schema/charities.ts
import {
  pgTable,
  serial,
  text,
  boolean,
  date,
  timestamp,
  numeric,
} from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const charitiesTable = pgTable("charities", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  description: text("description").notNull(),
  imageUrl: text("image_url"),
  website: text("website"),
  category: text("category"),
  featured: boolean("featured").notNull().default(false),
  upcomingEvent: text("upcoming_event"),
  upcomingEventDate: date("upcoming_event_date"),
  totalReceived: numeric("total_received", { precision: 12, scale: 2 })
    .notNull()
    .default("0"),
  createdAt: timestamp("created_at", { withTimezone: true })
    .notNull()
    .defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true })
    .notNull()
    .defaultNow()
    .$onUpdate(() => new Date()),
});

export const insertCharitySchema = createInsertSchema(charitiesTable).omit({
  id: true,
  createdAt: true,
  updatedAt: true,
});
export type InsertCharity = z.infer<typeof insertCharitySchema>;
export type Charity = typeof charitiesTable.$inferSelect;